+++ avcodec/packet.go
@@ -53,17 +53,17 @@ func AvPacketFree(p *Packet) {
 
 //Allocate new information of a packet.
 func (p *Packet) AvPacketNewSideData(t AvPacketSideDataType, s int) *uint8 {
-	return (*uint8)(C.av_packet_new_side_data((*C.struct_AVPacket)(p), (C.enum_AVPacketSideDataType)(t), C.int(s)))
+	return (*uint8)(C.av_packet_new_side_data((*C.struct_AVPacket)(p), (C.enum_AVPacketSideDataType)(t), C.ulong(s)))
 }
 
 //Shrink the already allocated side data buffer.
 func (p *Packet) AvPacketShrinkSideData(t AvPacketSideDataType, s int) int {
-	return int(C.av_packet_shrink_side_data((*C.struct_AVPacket)(p), (C.enum_AVPacketSideDataType)(t), C.int(s)))
+	return int(C.av_packet_shrink_side_data((*C.struct_AVPacket)(p), (C.enum_AVPacketSideDataType)(t), C.ulong(s)))
 }
 
 //Get side information from packet.
 func (p *Packet) AvPacketGetSideData(t AvPacketSideDataType, s *int) *uint8 {
-	return (*uint8)(C.av_packet_get_side_data((*C.struct_AVPacket)(p), (C.enum_AVPacketSideDataType)(t), (*C.int)(unsafe.Pointer(s))))
+	return (*uint8)(C.av_packet_get_side_data((*C.struct_AVPacket)(p), (C.enum_AVPacketSideDataType)(t), (*C.ulong)(unsafe.Pointer(s))))
 }
 
 //Convenience function to free all the side data stored.
